Re-work object store de-duplication for proper typing.
